Refactors code to appropriately named specific controllers
