+++ src/SFA.DAS.Tools.Support.Web/Views/Support/Index.cshtml
@@ -16,7 +16,7 @@
                     This tool sends a pause command to approvals to update the commitments database and subsequently payments.
                 </dd>
                 <dd class="govuk-summary-list__actions">
-                    <a asp-controller="Approvals" asp-action="SearchApprenticeships" asp-route-act="pause" class="govuk-link">
+                    <a asp-controller="SearchApprovals" asp-action="SearchApprenticeships" asp-route-act="pause" class="govuk-link">
                         Open<span class="govuk-visually-hidden"> Pause apprenticeship(s)</span>
                     </a>
                 </dd>
@@ -29,7 +29,7 @@
                     This tool sends a resume command to approvals to update the commitments database and subsequently payments.
                 </dd>
                 <dd class="govuk-summary-list__actions">
-                    <a asp-controller="Approvals" asp-action="SearchApprenticeships" asp-route-act="resume" class="govuk-link">
+                    <a asp-controller="SearchApprovals" asp-action="SearchApprenticeships" asp-route-act="resume" class="govuk-link">
                         Open<span class="govuk-visually-hidden"> Resume apprenticeship(s)</span>
                     </a>
                 </dd>
@@ -42,7 +42,7 @@
                     This tool sends a stop command to approvals to update the commitments database and subsequently payments.
                 </dd>
                 <dd class="govuk-summary-list__actions">
-                    <a asp-controller="Approvals" asp-action="SearchApprenticeships" asp-route-act="stop" class="govuk-link">
+                    <a asp-controller="SearchApprovals" asp-action="SearchApprenticeships" asp-route-act="stop" class="govuk-link">
                         Open<span class="govuk-visually-hidden"> Stop apprenticeship(s)</span>
                     </a>
                 </dd>
